Refactor the code to reduce redundant code
